requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
astro-sdk-python[amazon,google,snowflake,mssql,postgres]